+++ tests/testthat/test-fetch-ladder.R
@@ -10,8 +10,17 @@ test_that("fetch_ladder_afl works for various inputs", {
   expect_error(suppressWarnings(fetch_ladder_afl(2000, 1, comp = "AFLM")))
 
   # change round number
-  expect_s3_class(fetch_ladder_afl(2020, round_number = 2), "tbl")
-  expect_s3_class(fetch_ladder_afl(2020, round_number = 50), "tbl")
+  lad <- fetch_ladder_afl(2020, round_number = 2)
+  expect_s3_class(lad, "tbl")
+  expect_equal(max(lad$round_number), 2)
+  expect_equal(min(lad$round_number), 2)
+  
+  lad <- fetch_ladder_afl(2021, comp = "AFLW", round_number = 5)
+  expect_s3_class(lad, "tbl")
+  expect_equal(max(lad$round_number), 5)
+  expect_equal(min(lad$round_number), 5)
+  
+  expect_error(fetch_ladder_afl(2020, round_number = 50))
 
   # change comp
   expect_s3_class(fetch_ladder_afl(2020, round_number = 1, comp = "AFLW"), "tbl")